+++ strato/VM/SolidVM/solid-vm/src/Blockchain/SolidVM.hs
@@ -1356,7 +1356,7 @@ runStatement st@(CC.EmitStatement eventName exptups pos) = do
           -- pair up field names with values one-by-one (no type checking tho, lol)
           -- let pairs = zip (map (T.unpack . fst) $ CC._eventLogs ev) expStrs
 
-          let evArgs = zipWith (\(name, CC.IndexedType _ idxType) value -> 
+          let evArgs = zipWith (\(CC.EventLog name _ (CC.IndexedType _ idxType)) value -> 
                         (T.unpack name, value, if isTypeArray idxType then "Array" else "Other")) 
                      (CC._eventLogs ev) expStrs
                 where